+++ NEWS.md
@@ -7,3 +7,4 @@
 * Added the axis scales `scale_x_nfl()` and `scale_y_nfl()` in combination with the theme update functions `theme_x_nfl()` and `theme_y_nfl()`. (v0.0.0.9004)
 * Fixed an incompatible position argument in `scale_y_nfl()`. (v.0.0.9005)
 * Added the function `ggpreview()` which allows to preview a ggplot in it's actual dimensions. (v.0.0.9006)
+* `geom_nfl_logos()` now supports a `colour` aesthetic that colorizes the logos. (v0.0.9007)